+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+
+## [0.29.1] - 2022-05-02
+* fix: forward mouse events to plugin panes (https://github.com/zellij-org/zellij/pull/1369)
+
+## [0.29.0] - 2022-05-02
 * add: clarify copy to clipboard message (https://github.com/zellij-org/zellij/pull/1321)
 * Terminal compatibility: fix ANSI scrolling regression (https://github.com/zellij-org/zellij/pull/1324)
 * fix: send SIGHUP instead of SIGTERM when closing a pane (https://github.com/zellij-org/zellij/pull/1320)
@@ -13,6 +18,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * add: add command for auto-start script (https://github.com/zellij-org/zellij/pull/1281)
 * Terminal compatibility: fix cursor pane escape and invalid ansi crash (https://github.com/zellij-org/zellij/pull/1349)
 * fix: recover from corrupted ipc bus state (https://github.com/zellij-org/zellij/pull/1351)
+* Terminal compatibility: respond to foreground/background color ansi requests (OSC 10 and 11) (https://github.com/zellij-org/zellij/pull/1358)
+* fix: avoid panic in link_handler.rs (https://github.com/zellij-org/zellij/pull/1356)
+* Terminal compatibility: prevent wide chars from overflowing the title line (https://github.com/zellij-org/zellij/pull/1361)
+* Terminal compatibility: adjust saved cursor position on resize (https://github.com/zellij-org/zellij/pull/1362)
+* fix: avoid panic on renaming a floating pane (https://github.com/zellij-org/zellij/pull/1357)
+* fix: change the way sessions are sorted (https://github.com/zellij-org/zellij/pull/1347)
+* fix: improve mouse event reporting, avoid clicks on plugin panes causing active pane scrolling (https://github.com/zellij-org/zellij/pull/1329)
 
 ## [0.28.1] - 2022-04-13
 * (BREAKING CHANGE) Feature: Improve theme usage and add default themes. Remove gray color from themes. (https://github.com/zellij-org/zellij/pull/1274)